fix: Modify Font and Texture classes
Since each line of text we try to print requires a dedicated texture, we
modified the Font class to become just a carrier of the TTF_Font
pointer. The rendering will still be carried by the Texture class, which
was modified to create SDL_Texture objects from rendered text. We also
modified the Texture class Rule of 5 for avoiding pointers conflict when
using Copy constructor or Copy assignment operator.